requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
WTForms==2.3.3
Request